+++ truffle/src/com.oracle.truffle.dsl.processor/src/com/oracle/truffle/dsl/processor/parser/NodeParser.java
@@ -2983,6 +2983,21 @@ public final class NodeParser extends AbstractParser<NodeData> {
             assumptionId++;
         }
         specialization.setAssumptionExpressions(assumptionExpressions);
+
+        if (!assumptionExpressions.isEmpty() && !specialization.isGuardBindsCache()) {
+            specialization.addSuppressableWarning(TruffleSuppressedWarnings.ASSUMPTION, """
+                            It is discouraged to use the assumptions property with a specialization that cannot have multiple specialization instances.\s\
+                            Note that assumption specializations get removed if the assumption is no longer true.\s\
+                            This rarely makes sense for specializations without multiple instances.\s\
+                            Most often this can be fixed by translating the assumption usage to a regular method guard instead.\s\
+                            So instead of assumptions="a" you may use guards="a.isValid()".\s\
+                            If your specialization should have multiple instances, make sure you bind a cached value in the guard.\s\
+                            """,
+                            getSimpleName(types.Fallback),
+                            getSimpleName(types.Specialization),
+                            getSimpleName(types.Fallback));
+        }
+
     }
 
     private void initializeLimit(SpecializationData specialization, DSLExpressionResolver resolver, boolean uncached) {